+++ src/Math/vector/Vector.java
@@ -4,23 +4,23 @@ public interface Vector {
 
     Vector vectorProduct(Vector v1, Vector v2);
 
-    void setValues(double[] values);
+    void setValues(float[] values);
 
     Vector plusTwoVectors(Vector v1, Vector v2);
 
     Vector minusTwoVectors(Vector v1, Vector v2);
 
-    Vector multiplyVectorOnScalar(Vector v1, double scalar);
+    Vector multiplyVectorOnScalar(Vector v1, float scalar);
 
-    Vector divisionVectorOnScalar(Vector v1, double scalar);
+    Vector divisionVectorOnScalar(Vector v1, float scalar);
 
-    double vectorLength(Vector v1);
+    float vectorLength(Vector v1);
 
     Vector vectorNormalization(Vector v1);
 
-    double scalarMultiplyVectorOnVector(Vector v1, Vector v2);
+    float scalarMultiplyVectorOnVector(Vector v1, Vector v2);
 
-    double[] getValues();
+    float[] getValues();
 
     int getSize();
